+++ packages/moonstone/Scrollable/Scrollable.js
@@ -448,6 +448,10 @@ class ScrollableBase extends Component {
 
 		forward('onKeyDown', ev, this.props);
 
+		if (isPageUp(keyCode) || isPageDown(keyCode)) {
+			ev.preventDefault();
+		}
+
 		this.animateOnFocus = true;
 
 		if (!repeat && this.hasFocus()) {
@@ -457,7 +461,6 @@ class ScrollableBase extends Component {
 				direction = null;
 
 			if (isPageUp(keyCode) || isPageDown(keyCode)) {
-				ev.preventDefault();
 				if (this.props.direction === 'vertical' || this.props.direction === 'both') {
 					Spotlight.setPointerMode(false);
 					direction = isPageUp(keyCode) ? 'up' : 'down';